app/requirements.txt
flask
psycopg2-binary
flake8